+++ src/main/Anvil/Services/Services/BindingOrder.cs
@@ -1,9 +1,0 @@
-namespace Anvil.Services
-{
-  internal enum BindingOrder : short
-  {
-    Core = short.MinValue,
-    API = -30000,
-    Default = 15000,
-  }
-}